project/backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
pydantic>=1.10.0
PyJWT>=2.8.0
python-multipart>=0.0.9
pymysql>=1.1.0
datasets>=2.18.0
pyarrow>=14.0.0
pandas>=2.1.0
av>=10.0.0
Pillow>=10.0.0
